Don't flag all political sources as unreliable
Many websites under the Political category are biased towards a certain political orientation or view, however they provide mostly factual news, so they shouldn't be marked as "unreliable", but rather as:

1) : "may or may not be reliable" - The same way websites in the "Unknown" category are marked.

or

2) "reliable" - The same way websites in the "Reliable" category are marked.

At the time being, Political websites will be marked as (1). The database provides additional tags for many websites, so in the next version, the app will use them to display websites that provide high factual reporting as "Reliable".